Move auth providers to a workspace package
