+++ packages/care-ops-auth/package.json
@@ -1,14 +1,26 @@
 {
   "private": true,
-  "name": "@roundingwell/care-ops-config",
+  "name": "@roundingwell/care-ops-auth",
   "version": "1.0.0",
-  "description": "Config util for Care‑Ops Frontend",
+  "description": "Auth providers for Care‑Ops Frontend",
   "author": "RoundingWell",
   "license": "CC-BY-4.0",
   "homepage": "https://roundingwell.com",
   "type": "module",
-  "files": ["index.js"],
-  "main": "./index.js",
-  "exports": "./index.js",
-  "sideEffects": false
+  "files": [
+    "AuthProvider.js",
+    "providers/"
+  ],
+  "exports": {
+    "./AuthProvider.js": "./AuthProvider.js",
+    "./auth0.js": "./providers/auth0.js",
+    "./kinde.js": "./providers/kinde.js",
+    "./workos.js": "./providers/workos.js"
+  },
+  "sideEffects": false,
+  "dependencies": {
+    "@auth0/auth0-spa-js": "^2.1.3",
+    "@kinde-oss/kinde-auth-pkce-js": "^4.2.0",
+    "@workos-inc/authkit-js": "^0.12.0"
+  }
 }
