+++ packages/care-ops-auth/providers/kinde.js
@@ -1,14 +1,14 @@
-import { AuthProvider } from './AuthProvider.js';
+import { AuthProvider } from '../AuthProvider.js';
 
-import { createClient } from '@workos-inc/authkit-js';
+import createKindeClient from '@kinde-oss/kinde-auth-pkce-js';
 
-export class WorkosAuthProvider extends AuthProvider {
+export class KindeAuthProvider extends AuthProvider {
   async getToken() {
     if (!this.client) return;
     if (!navigator.onLine && this.token) return this.token;
 
     return this.client
-      .getAccessToken()
+      .getToken()
       .then(token => {
         this.token = `Bearer ${ token }`;
         return this.token;
@@ -19,30 +19,22 @@ export class WorkosAuthProvider extends AuthProvider {
       });
   }
 
-  login(path = AuthProvider.PATH_ROOT) {
-    this.client.signIn({ state: path });
+  login(path, connection = this.config.connections.default) {
+    this.client.register({
+      app_state: { path },
+      authUrlParams: { connection_id: connection },
+    });
   }
 
-  // If considered RW and rwClientId is set
-  _getClientId(pathName) {
-    const { clientId, rwClientId } = this.config;
-
-    // RWell specific login
-    if (rwClientId && (pathName === AuthProvider.PATH_RWELL || localStorage.getItem(AuthProvider.PATH_RWELL))) {
-      return rwClientId;
-    }
-
-    return clientId;
-  }
-
-  async _initClient(clientId) {
+  async _initClient() {
     return new Promise(resolve => {
       const clientConfig = {
-        redirectUri: location.origin + AuthProvider.PATH_AUTHD,
-        onRedirectCallback: ({ user, state }) => {
-          const path = state;
+        ...this.config.createParams,
+        redirect_uri: location.origin + AuthProvider.PATH_AUTHD,
+        logout_uri: location.origin,
+        on_redirect_callback: (user, { path } = {}) => {
           if (!user) {
-            this.loginPrompt(path);
+            this.loginPrompt({ appState: path });
             return;
           }
 
@@ -50,10 +42,9 @@ export class WorkosAuthProvider extends AuthProvider {
 
           resolve();
         },
-        ...this.config.createClientOptions,
       };
 
-      createClient(clientId, clientConfig)
+      createKindeClient(clientConfig)
         .then(client => {
           this.client = client;
         });
@@ -67,15 +58,19 @@ export class WorkosAuthProvider extends AuthProvider {
 
     const pathName = location.pathname;
 
-    const clientId = this._getClientId(pathName);
-
-    await this._initClient(clientId);
+    await this._initClient();
 
     if (pathName === AuthProvider.PATH_AUTHD) return;
 
     if (pathName === AuthProvider.PATH_LOGOUT) {
       this.token = null;
-      this.client.signOut({ returnTo: location.origin });
+      this.client.logout();
+      return;
+    }
+
+    // RWell specific login
+    if (pathName === AuthProvider.PATH_RWELL || localStorage.getItem(AuthProvider.PATH_RWELL)) {
+      this.login(AuthProvider.PATH_RWELL, this.config.connections.roundingwell);
       return;
     }
